refactor: refactor interfaces and modules
