+++ src/components/index.ts
@@ -1,6 +1,6 @@
 export { default as UserCard } from './_UserCard.svelte';
 
-export { default as CostsOverviewTripDetails } from './tripDetails/_CostsOverviewTripDetails.svelte'
+export { default as CostsOverviewTripDetails } from './tripDetails/_CostsOverviewTripDetails.svelte';
 export { default as TripDetailsHeader } from './tripDetails/_HeaderTripDetails.svelte';
 export { default as HeaderTripDetailsLayout } from './tripDetails/_HeaderTripDetailsLayout.svelte';
 export { default as TripDetailsCostOverview } from './tripDetails/_CostsOverviewTripDetails.svelte';
