+++ src/components/tripOverview/_DonutChart.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { Doughnut } from 'svelte-chartjs';
 	import { Chart as ChartJS, Title, Tooltip, ArcElement } from 'chart.js';
-	export let data: import('../../interfaces/TripOverview').ChartData;
+	export let data: import('../../interfaces/Trips').ChartData;
 
 	ChartJS.register(Title, Tooltip, ArcElement);
 </script>
